+++ examples/python/qb_mps_noisy.py
@@ -21,9 +21,9 @@ s.measure_sample_sequential = "auto"
 
 # Uncomment the following lines to introduce noise to the simulation.
 # You will need to have the Qristal emulator installed for this to work.
-#s.noise = True
-#nm = qristal.core.NoiseModel("qb-nm1", n_qbits)
-#s.noise_model = nm
+# s.noise = True
+# nm = qristal.core.NoiseModel("qb-nm1", n_qbits)
+# s.noise_model = nm
 
 # In this test we use cx to perform a standard
 # cx on |111>. Expected outcome: |011>
@@ -42,5 +42,20 @@ circ.measure_all()
 
 # run the circuit and check results
 s.ir_target = circ
+
+# CudaQ has no transpiler, so we need to transpile the circuit to QB's native gate set {rx, ry, cz} first.
+if s.acc[0][0] == "cudaq:qb_mps" and s.noise[0][0] == True:
+  # To do this, we simply run the program without executing the circuit, i.e. by setting nosim = True
+  s.nosim = True
+  s.run()
+
+  # Now we can get the transpiled circuit using "out_transpiled_circuit"
+  circ_qasm = s.out_transpiled_circuit[0][0]
+  # print(circ_qasm)
+  # The transpiled circuit is in openQasm, so we feed it back into session via "instring"
+  s.instring = circ_qasm
+
+  # Execute the transpiled circuit by setting nosim = False
+  s.nosim = False
 s.run()
 print(s.results[0][0])